Add currtime print at end of test to try and figure out why ALRM is going off...
git-svn-id: https://svn.internet2.edu/svn/bwctl/trunk@94 2da7661c-602c-0410-9277-d9994c0c2ef1

+++ bwlib/endpoint.c
@@ -945,6 +945,16 @@ end:
 		BWLError(ctx,BWLErrFATAL,errno,"setitimer(): %M");
 	}
 
+	if(!BWLGetTimeStamp(ctx,&currtime)){
+		BWLError(ctx,BWLErrFATAL,BWLErrUNKNOWN,
+				"BWLGetTimeStamp(): %M");
+		abort();
+	}
+	BWLError(ctx,BWLErrINFO,BWLErrINVALID,
+			"End of test: currtime = %f",
+			BWLNum64ToDouble(currtime.tstamp)
+			);
+
 	if((kill(ep->child,SIGINT) != 0) && (errno != ESRCH)){
 		BWLError(ctx,BWLErrFATAL,errno,
 				"Unable to kill test endpoint, pid=%d: %M",
